pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "helio"
version = "0.3.0"
description = "AI Companion for Solar PV Simulation using pvlib"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Fiacre Rougieux", email = "fiacrerougieux@gmail.com"}
]
keywords = ["solar", "pv", "pvlib", "simulation", "ai", "llm", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pvlib>=0.14.0,<0.15.0",
    "numpy>=1.24.0,<2.0.0",
    "pandas>=2.0.0,<3.0.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "keyring>=24.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
training = [
    # These are large packages, make them optional
    # "torch>=2.1.0",
    # "transformers>=4.36.0",
    # "datasets>=2.14.0",
    # "trl>=0.7.0",
]

[project.scripts]
helio = "agent.multi_agent_cli:main"
helio-auth = "agent.auth_cli:main"
sun = "agent.multi_agent_cli:main"

[project.urls]
Homepage = "https://github.com/fiacrerougieux/sun-sleuth-dev"
Repository = "https://github.com/fiacrerougieux/sun-sleuth-dev"

[tool.setuptools]
packages = ["agent", "training"]

[tool.setuptools.package-data]
agent = ["*.py"]
training = ["*.py"]
